+++ NEWS.md
@@ -5,6 +5,7 @@
 * Update "The Bank Tutorial: Part I" vignette (@nacnudus in #90).
 * Fix `trap()`'s handler cloning and associated test (#91).
 * Apply `select()`'s `policy` also when `resources` is a function (#92).
+* Accept dynamic timeouts in batches (#93).
 
 # simmer 3.6.1