+++ NEWS.md
@@ -6,6 +6,7 @@
 * Fix `trap()`'s handler cloning and associated test (#91).
 * Apply `select()`'s `policy` also when `resources` is a function (#92).
 * Accept dynamic timeouts in batches (#93).
+* Change `rollback()`'s default behaviour to `times=Inf`, i.e., infinite loop (#95).
 
 # simmer 3.6.1