+++ NEWS.md
@@ -1,4 +1,4 @@
-# simmer 3.6.1.9000
+# simmer 3.6.2
 
 ## Minor changes and fixes:
 
@@ -8,8 +8,8 @@
 * Accept dynamic timeouts in batches (#93).
 * Change `rollback()`'s default behaviour to `times=Inf`, i.e., infinite loop (#95).
 * Stop and throw an error when `timeout()` returns a missing value (#96 and #97).
-* Fix memory management: resetting the environment was clearing but not deallocating memory (#98).
-* Fix object destruction: workaround for tidyverse/magrittr#146.
+* Fix memory management: resetting the environment was clearing but not deallocating memory (#98, fixed in #99).
+* Fix object destruction: workaround for tidyverse/magrittr#146 (#98, fixed in effcb6b).
 
 # simmer 3.6.1